radiology-grpo-2/radiology-grpo-2/requirements.txt
torch>=2.0.0
transformers>=4.40.0
numpy>=1.23.0
tqdm>=4.66.0
torchvision>=0.16.0
Pillow>=10.0.0
evaluate
rouge-score
matplotlib